legacy/requirements.txt
# ── Web Server ─────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── MQTT ───────────────────────────────────────
paho-mqtt>=2.1.0

# ── Config ─────────────────────────────────────
python-dotenv>=1.0.0
